refactor: formatTaskAsTicket を formatter.go に分離
list.go に混在していた表示ヘルパーを cmd/formatter.go に切り出し
責務を明確化
