+++ cmd/list.go
@@ -8,16 +8,14 @@ import (
 	"log"
 
 	"github.com/Sottiki/docketpunch/internal/storage"
-	"github.com/Sottiki/docketpunch/internal/task"
 	"github.com/spf13/cobra"
 )
 
-// listCmd represents the list command
 var listCmd = &cobra.Command{
 	Use:   "list",
 	Short: "List all tasks",
 	Long: `List all tasks in ticket format.
-	
+
 Example:
   docket list`,
 	Run: func(cmd *cobra.Command, args []string) {
@@ -39,24 +37,4 @@ Example:
 
 func init() {
 	rootCmd.AddCommand(listCmd)
-}
-
-// タスクをチケット形式で表示する
-func formatTaskAsTicket(t *task.Task) string {
-	statusMark := " "
-	if t.Done {
-		statusMark = "◯"
-	}
-
-	createDate := t.CreatedAt.Format("01/02")
-
-	var dateInfo string
-	if t.Done && t.CompletedAt != nil {
-		completeDate := t.CompletedAt.Format("01/02")
-		dateInfo = fmt.Sprintf("(%s→%s)", createDate, completeDate)
-	} else {
-		dateInfo = fmt.Sprintf("(%s)", createDate)
-	}
-	return fmt.Sprintf("[ %s|#%d|%s %s]", statusMark, t.ID, t.Description, dateInfo)
-
 }
